+++ elijah-core/src/main/java/tripleo/elijah/comp/internal/CCI_Acceptor__CompilerInputListener.java
@@ -4,7 +4,6 @@ import tripleo.elijah.*;
 import tripleo.elijah.ci.*;
 import tripleo.elijah.comp.*;
 import tripleo.elijah.comp.i.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.util.*;
 
 import java.util.*;
