+++ elijah-core/src/main/java/tripleo/elijah/comp/nextgen/CX_ParseElijahFile.java
@@ -10,7 +10,6 @@ import tripleo.elijah.comp.internal.*;
 import tripleo.elijah.comp.specs.*;
 import tripleo.elijah.diagnostic.*;
 import tripleo.elijah.lang.i.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.util.*;
 import tripleo.elijjah.*;
