+++ elijah-core/src/main/java/tripleo/elijah/lang/types/OS_UserClassType.java
@@ -2,7 +2,6 @@ package tripleo.elijah.lang.types;
 
 import org.jetbrains.annotations.*;
 import tripleo.elijah.lang.i.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.stages.deduce.*;
 import tripleo.elijah.stages.gen_fn.*;
 import tripleo.elijah.util.*;
