+++ elijah-core/src/main/java/tripleo/elijah/nextgen/inputtree/EIT_Input_HashSourceFile_Triple.java
@@ -2,7 +2,7 @@ package tripleo.elijah.nextgen.inputtree;
 
 import org.jetbrains.annotations.NotNull;
 import tripleo.elijah.comp.*;
-import tripleo.elijah.nextgen.query.Mode;
+import tripleo.elijah.util.Mode;
 import tripleo.elijah.util.Helpers;
 import tripleo.elijah.util.Operation;
