+++ elijah-core/src/main/java/tripleo/elijah/sense/SenseList.java
@@ -4,7 +4,6 @@ import org.jetbrains.annotations.*;
 import tripleo.elijah.ci.*;
 import tripleo.elijah.comp.*;
 import tripleo.elijah.comp.i.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.util.*;
 
 import java.util.*;
