+++ elijah-core/src/main/java/tripleo/elijah/stages/deduce/Unnamed_ITE_Resolver1.java
@@ -5,7 +5,6 @@ import org.jetbrains.annotations.*;
 import tripleo.elijah.lang.i.*;
 import tripleo.elijah.lang.impl.*;
 import tripleo.elijah.lang.types.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.stages.gen_fn.*;
 import tripleo.elijah.stages.gen_fn.IdentTableEntry.*;
 import tripleo.elijah.stages.instructions.*;
