+++ elijah-core/src/main/java/tripleo/elijah/stages/deduce/post_bytecode/DeduceElement3_IdentTableEntry.java
@@ -3,11 +3,9 @@ package tripleo.elijah.stages.deduce.post_bytecode;
 import org.jdeferred2.*;
 import org.jetbrains.annotations.*;
 import tripleo.elijah.contexts.*;
-import tripleo.elijah.diagnostic.*;
 import tripleo.elijah.lang.i.*;
 import tripleo.elijah.lang.impl.*;
 import tripleo.elijah.lang.imports.*;
-import tripleo.elijah.nextgen.query.*;
 import tripleo.elijah.stages.deduce.*;
 import tripleo.elijah.stages.deduce.nextgen.*;
 import tripleo.elijah.stages.deduce.post_bytecode.DED.*;
