+++ elijah-core/src/main/java/tripleo/elijah/stages/gen_fn/EvaClass.java
@@ -16,7 +16,7 @@ import tripleo.elijah.lang.types.OS_BuiltinType;
 import tripleo.elijah.lang.types.OS_GenericTypeNameType;
 import tripleo.elijah.lang.types.OS_UnknownType;
 import tripleo.elijah.lang.types.OS_UserClassType;
-import tripleo.elijah.nextgen.query.Mode;
+import tripleo.elijah.util.Mode;
 import tripleo.elijah.nextgen.reactive.DefaultReactive;
 import tripleo.elijah.nextgen.reactive.Reactive;
 import tripleo.elijah.stages.deduce.*;
@@ -28,7 +28,6 @@ import tripleo.elijah.util.Helpers;
 import tripleo.elijah.util.NotImplementedException;
 import tripleo.elijah.util.Operation;
 import tripleo.elijah.world.i.LivingClass;
-import tripleo.elijah.world.impl.DefaultLivingClass;
 
 import java.util.*;
 import java.util.function.Consumer;
