+++ elijah-core/src/main/java/tripleo/elijah/stages/gen_fn/EvaContainer.java
@@ -16,7 +16,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import tripleo.elijah.lang.i.*;
 import tripleo.elijah.lang.types.OS_UserType;
-import tripleo.elijah.nextgen.query.Mode;
+import tripleo.elijah.util.Mode;
 import tripleo.elijah.stages.deduce.DeduceTypes2;
 import tripleo.elijah.stages.deduce.RegisterClassInvocation_env;
 import tripleo.elijah.stages.deduce.post_bytecode.DeduceElement3_VarTableEntry;
