+++ elijah-core/src/main/java/tripleo/elijah/util/Mode.java
@@ -1,4 +1,4 @@
-package tripleo.elijah.nextgen.query;
+package tripleo.elijah.util;
 
 public enum Mode {
 	FAILURE, NOTHING, SUCCESS
